adds and implements RealmService and fixes small bugs
